fix S11P12D203-307 : 점령지 상세화면 조회 수정

+++ backend/mallang/src/main/java/com/chill/mallang/domain/faction/dto/FactionDTO.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @Builder
 public class FactionDTO {
     private Long teamId;
-    private int teamPoint;
+    private Float teamPoint;
     private TopUserDTO topUser;
 }
